C#: Reorganize patterns
